# Kitty: use deflate for initial upload when animating