--- a/src/demo/xray.c
+++ b/src/demo/xray.c
@@ -180,7 +180,7 @@ int xray_demo(struct notcurses* nc){
   marsh.next_frame = 0;
   marsh.last_frame_rendered = -1;
   marsh.lplane = NULL;
-  marsh.dm = notcurses_check_pixel_support(nc) < 1 ? 0.5 * delaymultiplier : 0;
+  marsh.dm = notcurses_check_pixel_support(nc) ? 0 : 0.5 * delaymultiplier;
   int ret = -1;
   // FIXME need do something about SIGINT here, which can leave us locked up
   if(pthread_create(&tid1, NULL, xray_thread, NULL)){