… NEWS.md
… tools/release.sh