When we're using Kitty animation, we don't need to keep the encoded glyph around and edit it in-place, so feed that fucker through `deflate()` from zlib, using the `o=z` control. For `xray` (admittedly something of a favourable example), this reduces bandwidth by 96%(!). Unfortunately, it's kinda slow right now; the bandwidth savings only yield a net positive when operating over a reasonably slow network like my local wireless. I think we can improve our usage of zlib and make it faster, though.

```c
Samples: 113K of event 'cycles', Event count (approx.): 116203165085
  Children      Self  Command         Shared Object                  Symbol
    28.29%    28.25%  notcurses-demo  libswscale.so.5.7.100          [.] yuv2rgba32_full_X_c
-   16.80%     0.00%  notcurses-demo  [unknown]                      [k] 0000000000000000
   - 0
        7.96% kitty_blit_core
      - 0.99% 0x7f84496ea0b0
           0.98% ff_hscale14to15_X4_ssse3.innerloop
      - 0.76% 0x7f84496ea018
           0.58% rgbaToA_c
-   12.32%    12.30%  notcurses-demo  libnotcurses-core.so.2.3.11    [.] kitty_blit_core
   - 7.95% 0
        kitty_blit_core
     1.72% kitty_blit_core
   - 1.09% 0x24900000000
        kitty_blit_core
-   10.70%    10.68%  notcurses-demo  libz.so.1.2.11                 [.] deflate_slow
     10.14% deflate_slow
-    8.59%     8.58%  notcurses-demo  libswscale.so.5.7.100          [.] ff_hscale14to15_X4_ssse3.inner
   - 4.33% 0
      - 0.98% 0x7f84496ea0b0
           ff_hscale14to15_X4_ssse3.innerloop
   - 4.24% 0x616c665f73777300
        sws_context_to_name
        swscale
        ff_hscale14to15_X4_ssse3.innerloop
-    7.84%     7.83%  notcurses-demo  libz.so.1.2.11                 [.] fill_window
     fill_window
-    6.77%     6.75%  notcurses-demo  libz.so.1.2.11                 [.] longest_match
     longest_match
+    4.34%     0.03%  notcurses-demo  libswscale.so.5.7.100          [.] swscale
+    4.34%     0.00%  notcurses-demo  [unknown]                      [.] 0x616c665f73777300
+    4.34%     0.00%  notcurses-demo  libswscale.so.5.7.100          [.] sws_context_to_name
+    2.82%     2.81%  notcurses-demo  libz.so.1.2.11                 [.] adler32_z
+    2.55%     2.54%  notcurses-demo  libswscale.so.5.7.100          [.] rgbaToA_c
+    2.44%     0.00%  notcurses-demo  libavutil.so.56.51.100         [.] av_default_item_name
```

that's a profile of `xray`. let's see what we can't claw back.

here's a run over my local wireless, showing the potential winnage:

before

```
             runtime│ frames│output(B)│    FPS│%r│%a│%w│TheoFPS║
══╤════════╤════════╪═══════╪═════════╪═══════╪══╪══╪══╪═══════╣
 1│    xray│  65.73s│    486│   1.47Gi│    7.4│ 0│ 0│76│   9.64║
══╧════════╧════════╪═══════╪═════════╪═══════╧══╧══╧══╧═══════╝
              65.73s│    486│   1.47Gi│
```

after

```
             runtime│ frames│output(B)│    FPS│%r│%a│%w│TheoFPS║
══╤════════╤════════╪═══════╪═════════╪═══════╪══╪══╪══╪═══════╣
 1│    xray│  34.11s│    486│  66.33Mi│   14.2│ 0│ 0│ 2│ 592.13║
══╧════════╧════════╪═══════╪═════════╪═══════╧══╧══╧══╧═══════╝
              34.11s│    486│  66.33Mi│
```

Closes #1695.