@dankamongmen 2021-07-21
Oh, this PR also does other things:

* Properly switch between `SPRIXEL_ANNIHILATED` and `SPRIXEL_ANNIHILATED_TRANS` on reload
* Properly carry `transcolor`-elided pixels into auxvec so they're not restored visibly
* Tighten error checks in `intro`
* Fix crash on error path with kitty
* Remove cell-pixel geometry from `blitterargs`; it's available from the sprixel
* Properly call `scrub_tam_boundaries` in iTerm2 and linux